tests

+++ internal/core/service/location/location.go
@@ -9,6 +9,7 @@ import (
 
 var (
 	ErrLocationNotFound = errors.New("location not found")
+	ErrUnAuthorized     = errors.New("unauthorized")
 )
 
 type Service struct {
@@ -32,9 +33,24 @@ func (s *Service) GetLocations(ctx context.Context, userID string, filter Filter
 }
 
 func (s *Service) UpdateLocation(ctx context.Context, location domain.Location) (domain.Location, error) {
+	loc, err := s.repo.GetLocation(ctx, location.Id)
+	if err != nil {
+		return domain.Location{}, ErrLocationNotFound
+	}
+	if loc.UserID != location.UserID {
+		return domain.Location{}, ErrUnAuthorized
+	}
 	return s.repo.UpdateLocation(ctx, location)
 }
 
-func (s *Service) DeleteLocation(ctx context.Context, id string) error {
+func (s *Service) DeleteLocation(ctx context.Context, id string, userID string) error {
+	loc, err := s.repo.GetLocation(ctx, id)
+	if err != nil {
+		return ErrLocationNotFound
+	}
+	if loc.UserID != userID {
+		return ErrUnAuthorized
+	}
+
 	return s.repo.DeleteLocation(ctx, id)
 }
